register event form integrated
